project/glossary_project/swagger_urls.py
from rest_framework.permissions import IsAdminUser
from drf_yasg.views import get_schema_view
from drf_yasg import openapi
from django.urls import path, re_path

schema_view = get_schema_view(
   openapi.Info(
      title="IT Glossary API",
      default_version='v1',
      description="Документация API для глоссария IT-специалиста",
      contact=openapi.Contact(email="sianko0608@gmail.com"),
   ),
   public=True,
   permission_classes=(IsAdminUser,),
)

urlpatterns = [
   re_path(r'^swagger(?P<format>\.json|\.yaml)$', schema_view.without_ui(cache_timeout=0), name='schema-json'),
   path('swagger/', schema_view.with_ui('swagger', cache_timeout=0), name='schema-swagger-ui'),
   path('redoc/', schema_view.with_ui('redoc', cache_timeout=0), name='schema-redoc'),
]